# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Roleplay Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List my chats
      responses:
        "200":
          description: Chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat with a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a chat with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithMessages"
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update chat (system prompt)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChatBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat and its history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chats/{id}/messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePair"
  /chats/{id}/messages/{msgId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: msgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Group Chats ──────────────────────────────────────────────────────────
...
  /dm:
    get:
      operationId: listDmConversations
      tags: [chats]
      summary: List direct message conversations
      responses:
        "200":
          description: DM conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DmConversation"
    post:
      operationId: sendDm
      tags: [chats]
      summary: Send a direct message to a persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDmBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmMessage"
  /dm/{personaId}:
    get:
      operationId: getDmMessages
      tags: [chats]
      summary: Get DM messages with a persona
      parameters:
        - name: personaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DmMessage"
components:
  schemas:
...
    Character:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: string
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        profilePicture:
          type: ["string", "null"]
        backgroundStory:
          type: string
        lore:
          type: string
        introMessage:
          type: string
        tags:
          type: array
          items:
            type: string
        chatCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, ownerUserId, name, age, personality, description, backgroundStory, lore, introMessage, tags, chatCount, createdAt, updatedAt]
...
    Persona:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: string
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        lore:
          type: string
        tags:
          type: array
          items:
            type: string
        isMain:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, ownerUserId, name, age, personality, description, lore, tags, isMain, createdAt]
...
    Chat:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        characterId:
          type: integer
        personaId:
          type: integer
        character:
          $ref: "#/components/schemas/Character"
        persona:
          $ref: "#/components/schemas/Persona"
        systemPrompt:
          type: ["string", "null"]
        modelId:
          type: ["integer", "null"]
        lastMessageAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, characterId, personaId, createdAt]
    CreateChatBody:
      type: object
      properties:
        characterId:
          type: integer
        personaId:
          type: integer
        modelId:
          type: ["integer", "null"]
        systemPrompt:
          type: ["string", "null"]
      required: [characterId, personaId]
    UpdateChatBody:
      type: object
      properties:
        systemPrompt:
          type: ["string", "null"]
        modelId:
          type: ["integer", "null"]
    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        role:
          type: string
        content:
          type: string
        modelUsed:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, chatId, role, content, createdAt]
    ChatWithMessages:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        characterId:
          type: integer
        personaId:
          type: integer
        character:
          $ref: "#/components/schemas/Character"
        persona:
          $ref: "#/components/schemas/Persona"
        systemPrompt:
          type: ["string", "null"]
        modelId:
          type: ["integer", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        lastMessageAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, characterId, personaId, messages, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        modelId:
          type: ["integer", "null"]
      required: [content]
    MessagePair:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"
        modelUsed:
          type: string
        blocked:
          type: boolean
        blockReason:
          type: ["string", "null"]
      required: [userMessage, assistantMessage, modelUsed, blocked]
...
    DmConversation:
      type: object
      properties:
        personaId:
          type: integer
        persona:
          $ref: "#/components/schemas/Persona"
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
      required: [personaId, unreadCount]
    SendDmBody:
      type: object
      properties:
        toPersonaId:
          type: integer
        fromPersonaId:
          type: integer
        content:
          type: string
      required: [toPersonaId, fromPersonaId, content]
    DmMessage:
      type: object
      properties:
        id:
          type: integer
        fromPersonaId:
          type: integer
        toPersonaId:
          type: integer
        fromPersona:
          $ref: "#/components/schemas/Persona"
        content:
          type: string
        createdAt:
          type: string
      required: [id, fromPersonaId, toPersonaId, content, createdAt]