# lib/api-zod/src/generated/types/updateChatBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roleplay Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateChatBody {
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  modelId?: number | null;
}